bots: Move service bot tests into separate file.
